backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
pinai-agent-sdk==0.1.0
python-dotenv==1.0.0
python-multipart==0.0.6
httpx==0.25.0 